3.4-django-testing/django_testing/requirements.txt
django~=5.0.6
djangorestframework~=3.15.1
django-filter~=24.2
psycopg2-binary

pytest~=8.2.0